feat(mockserver): improve API body content checking
To ensure that our tests catch errors in the API requests we are sending
to a backend, we need to check the contents in the actual request we are
sending. A simple approach is to check if the body of the request
Contains or NotContains certain data.

This change improves the ReturnResponse feature of our mockserver to
optionally check for expected or unexpected strings in the request body.