fix(api): Fix race condition
If sending the POST request results in an error, the Go routine issuing
the request closes the pipe and then returns the error. Closing the
causes the other Go routine to fail with:

    io: read/write on closed pipe

Which of those two errors is returned first and therefore the error
which is returned by errgroup.Wait() is subject to a race condition.