feat(execute): Support executing single fuzz test bundle
... without specifying the fuzz test name. Use this feature in `cifuzz
container run`.

This allows us to use `cifuzz container run` with Bazel. It did not work
before because to specify a Bazel fuzz test, the user specifies a Bazel
label like `//src:explore_me_fuzz_test`. During build, cifuzz calls
Bazel to convert this label into a canonical path which is then used as
the fuzz test identifier in the bundle metadata. If `cifuzz container
run` just passes on the fuzz test argument, `cifuzz execute` can't find
a fuzz test with that name in the metadata file, and it can't call Bazel
to convert it because it's running in a container which probably doesn't
have Bazel installed. There are two solutions to solve that:

1. The bundler returns the fuzz test identifier and `cifuzz container
   run` passest hat on to `cifuzz execute`
2. `cifuzz container run` tells `cifuzz execute` to execute the only
   fuzz test inside the bundle.

This commit implements 2.