fix(coverage): Fix CouldBeSandboxError not passed on
`error.Errorf("%v", err)` creates a new error instead of wrapping
the existing error. That caused the CouldBeSandboxError which was
returned to not be passed on, so the root command didn't print the
note.

Using `fmt.Errorf("%w", err)` instead fixed this, because it wraps the
error. Note that the `%w` specifier is only supported by `fmt.Errorf`,
not by `errors.Errorf`.