feat(buildlog): Log build to stderr if stderr is not a terminal
Don't redirect the build output to a file if the output is not a terminal.
The reason for redirecting the build output in the first place is to
avoid spamming the user's terminal with too verbose output. This is not
necessary if the output is not a terminal.

This also has the effect that the build output is printed when cifuzz is
running in a CI environment, which makes it easier to retrieve the output
than if it was redirected to a file.