fix(test): Skip UBSan test on Windows
Since yesterday (2023-08-24), the test keeps failing in our CI because
symbols needed by clang_rt.ubsan_standalone-x86_64.lib can't be found.
This seems to be a known issue with newer Visual Studio versions:

    https://github.com/microsoft/STL/issues/3568

I suspect that the Visual Studio installed on the GitHub runner has been
updated and that we can't do anything but wait until the issue was
fixed.