fix(bundler): Ensure that build output is printed
Since we introduced the opts.BuildStdout and opts.BuildStderr fields,
various tests don't print the build output anymore which makes it
hard to debug them when they fail. This commit ensures that the build
output is printed even if the caller did not explicitly set the
respective fields in the opts struct.