feat: Run fuzz container with same UID as host user
This ensures that all files created below bind-mounted directories are
created with the same UID as the user running cifuzz instead of root.

We achieve that by passing the UID via an environment variable (CIFUZZ_UID)
to the container and make `cifuzz execute` change the ownership of the
/cifuzz directory to that UID and then execute the same command again
with the specified UID via syscall.Setuid and syscall.Exec. Those are
syscalls are only supported on Unix but that should be fine because we
don't support executing Windows containers anyway.

Other methods we considered but did not work:
* Creating the user when building the container image (via commands in
  the Dockerfile). We can't do that because we don't know at build time
  the UID of the user who will execute the container (we know that when
  using `cifuzz container run` but not when using `cifuzz container
  remote-run`).
* Setting the user in the host config of the Docker container (i.e. the
  same as running a container with `docker run --user $UID`). This
  breaks because the /cifuzz directory is not owned by UID, so creating
  files and directories is not permitted.
* Using Docker's userns-remap feature. This requires adapting the config
  of the Docker daemon, which users might not want to do. It also seems
  to be broken on my system (Debian unstable), where it fails with a
  "error mounting "cgroup" to rootfs" error.

This commit also makes the `cifuzz container run` command mount the
current working directory into the container, to allow the fuzz
container to copy inputs into the seed corpus directories (it was
necessary to do that in the same commit to avoid test failures).