fix: Fix error returned by os.UserConfigDir() being ignored
The error returned by os.UserConfigDir() was ignored, which resulted
in an incorrect file path being used for the access tokens file.

We now return that error in the Get() and Set() methods of package
tockenstorage.

This commit also contains some refactoring to reduce the number of calls
to auth.GetToken() and tokenstorage.Get().