fix: Fix CMake build log file not printed on error
Since commit 41582934f336ca83098d49d82d2770c64f6078e3, the err variable
that is passed to

    defer func(err *error) {
            if *err != nil {
                    buildPrinter.StopOnError(log.BuildInProgressErrorMsg)
            } else {
                    buildPrinter.StopOnSuccess(log.BuildInProgressSuccessMsg, true)
            }
    }(&err)

is not set on an build error, because err was redeclared via:

    cBuildResults, err := builder.Build([]string{c.opts.fuzzTest})

To avoid subtle errors like this in the future, this commit extracts a
new runCmd._buildFuzzTest() method. If that method returns an error, the
build log file is printed.