fix: Keep the original ID when adding error details to finding
We used to replace the original ID with a regex, for example

  "id": "shift_exponent"

was replaced with

  "id": "undefined behaviou?r: .*",

It's confusing that the error ID of a finding changes when more error
details are attached to it. It also broke the
TestIntegration_CPPErrors/shift_exponent test if an access token existed
on the system.