chore(deps): Return to upstream rules_fuzzing
All the fixes and features from our fork of rules_fuzzing have been
merged upstream and are now included in the v0.4.0 release.

The only functional difference between the fork and upstream is that
there is a dedicated `asan-ubsan` sanitizer and the
`cc_engine_instrumentation` flag is not repeatable. cifuzz will guide
users to update their dependency on rules_fuzzing.

This is a breaking change for users of the `cifuzz-replay` config,
which have to copy & paste the updated `.bazelrc` config from the
docs.