fix: Fix warning message emoji width
Some emojis, including the warning sign emoji, are printed with an incorrect
width in some widely used terminals, including GNOME Terminal and the iterm2
Terminal on Mac. This has the effect that the first letter of our warning
messages is partly covered by the warning sign emoji, which gives a
really bad look.

We now use the bell emoji instead for warning messages, which is not affected
by that issue.

Closes CLI-1126