fix: trim/add 'projects/' prefix correctly
... because the `/v1` and `/v2` API sends/expects the `projects/` prefix
but the `/v3` API doesn't and the backend will phase out external
projects IDs soon. In preparation, it is better to trim the prefix from
every project name that we can (i.e., in api.ListProjects) and only add
it to the (soon legacy) requests that still require it.
This has previosly led to some confusing bugs and 404 API requests
because the prefix was incorrectly prefixed.