fix(reporthandler): Set the startedAt time when the printer is started
We use this timestamp to print the time since the fuzzer was started.
That's closed to the time we call the Start() method than the time we
initialize the UpdatingPrinter - for example for NodeJS projects, the
same command builds and runs the fuzz test, and we don't want the build
time to be counted as test run time.