test: Fix `container run` integration tests when using glibc >= 2.38
Our integration tests used ubuntu:22.04 as the base image which has
glibc 2.37 installed. That caused the tests to fail if the host has
glibc >= 2.38 installed because the fuzz test binaries are built on the
host and executed in the fuzz container.

As a workaround, we use ubuntu:rolling as the base image which currently
has glibc 2.38 installed.

Closes CLI-1337