chore: Don't force building with BuildKit
These images can be built without BuildKit and BuildKit is the default
since Docker 23, so we don't have to set DOCKER_BUILDKIT=1 anymore for
BuildKit to be used.

This allows us to build without BuildKit on a macOS GitHub Action runner
where Docker has been installed via the "setup-docker-on-macos" action
which doesn't support BuildKit.