ci: Cache Maven repository
This should speed up builds and tests in the CI significantly because
currently every time cifuzz is being built (which happens in each job
which runs tests or builds an installer), the Maven dependencies for
the list-fuzz-tests tool are being fetched.

Addional Maven dependencies are fetched for integration tests which test
Maven projects. Also see the code comment added by this commit.